feat:修复部分Bug

+++ base_function/src/main/java/fun/xb/basefunction/entity/count.java
@@ -6,9 +6,9 @@ import lombok.Data;
 public class count {
 
     //统计类型
-    Long type;
+    Integer type;
 
     //统计的记录的id
-    Long id;
+    Integer id;
 
 }
